requirements.txt
fastapi
uvicorn
python-multipart
streamlit
requests
qdrant-client
google-generativeai
sentence-transformers
python-dotenv
Pillow